requirements.txt
streamlit==1.30.0
Pillow==10.1.0
pandas==2.0.3
inflection==0.5.1 
folium==0.15.1
streamlit-folium==0.17.3
plotly-express==0.4.1